semana7/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cramjam==2.8.3
cycler==0.12.1
fastparquet==2024.5.0
fonttools==4.52.4
fsspec==2024.5.0
greenlet==3.0.3
idna==3.7
kiwisolver==1.4.5
lxml==5.2.2
matplotlib==3.9.0
numpy==1.26.4
packaging==24.0
pandas==2.0.3
pillow==10.3.0
psycopg2-binary==2.9.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrends==4.9.2
pytz==2024.1
requests==2.32.2
six==1.16.0
SQLAlchemy==1.4.52
sqlalchemy-redshift==0.8.14
tzdata==2024.1
urllib3==2.2.1
